Update OpenAPI schemas
This PR updates the OpenAPI schemas for authorization request so that it should
now be consistent with the code. In addition, the JSON schemas have been over-
hauled to reduce duplicate definitions and make them more restrictive and
readable.

Note the proposed naming scheme to distinguish between the different schemas.

## 📋 Checklist

* ✅ Lint checks passed on local machine.
* ✅ Unit tests passed on local machine.